Bugfix of slow ibarra algorithm
