+++ knapsack/README.md
