+++ knapsack/src/solution_dp.cpp
@@ -1,8 +1,6 @@
 #include <algorithm>
 #include <iostream>
-#include <numeric>
 #include <vector>
-#include <fstream>
 
 using namespace std;
 
@@ -12,8 +10,8 @@ vector<int> solve_knapsack(const vector<int>& c, const vector<int>& w, int max_w
     vector<int> dp(max_w + 1, -1);
     dp[0] = 0;
     for (int i = 0; i < n; ++i) {
-        for (int j = max_w; j >= 0; --j) {
-            if (j >= w[i] && dp[j - w[i]] != -1) {
+        for (int j = max_w; j >= w[i]; --j) {
+            if (dp[j - w[i]] != -1) {
                 dp[j] = max(dp[j], dp[j - w[i]] + c[i]);
             }
         }
@@ -53,60 +51,24 @@ vector<int> recover_answer(const vector<int>& c, const vector<int>& w, int c_opt
     throw 1;
 }
 
-pair<int, vector<int>> get_halh_approximation(const vector<int>& c, const vector<int>& w, int max_w) {
-    int n = c.size();
-    vector<int> id(n);
-    iota(id.begin(), id.end(), 0);
-    sort(id.begin(), id.end(), [&c, &w](int i, int j){
-       return c[i] * w[j] > c[j] * w[i];
-    });
-    int c_approx = 0;
-    int w_approx = 0;
-    int i = 0;
-    for (; i < n && w_approx + w[i] <= max_w; ++i) {
-        w_approx += w[i];
-        c_approx += c[i];
-    }
-    if (i < n) {
-        c_approx = max(c_approx, c[i]);
-    }
-    return { c_approx, vector<int>(id.begin(), id.begin() + i) };
-}
-
-pair<int, vector<int>> get_eps_approximation(const vector<int>& c, const vector<int>& w, int max_w, double eps) {
-    eps /= 2;
-    int n = c.size();
-
-    auto [c_half_approx, _] = get_halh_approximation(c, w, max_w);
-    int c_max = 2 * c_half_approx;
-    vector<int> c_scaled = c;
-    for (int& x : c_scaled) {
-        x = (x * 1. * n) / (c_max * eps);
-    }
-
-    auto dp = solve_knapsack(c, w, max_w);
-    int w_opt = max_element(dp.begin(), dp.end()) - dp.begin();
-    auto things = recover_answer(c, w, dp[w_opt], w_opt);
-    return { dp[w_opt], things };
-}
-
 int main() {
-    freopen("data/1.public", "r", stdin);
     ios_base::sync_with_stdio(false);
     cin.tie(nullptr);
     cout.tie(nullptr);
 
     int n, max_w;
     cin >> n >> max_w;
-    cout << n;
+
     vector<int> c(n), w(n);
     for (int i = 0; i < n; ++i) {
         cin >> c[i] >> w[i];
     }
 
-    auto [c_opt, things] = get_eps_approximation(c, w, max_w, 0.1);
+    auto dp = solve_knapsack(c, w, max_w);
+    int w_opt = max_element(dp.begin(), dp.end()) - dp.begin();
+    auto things = recover_answer(c, w, dp[w_opt], w_opt);
 
-    cout << c_opt << '\n';
+    cout << dp[w_opt] << '\n';
     for (int id : things) {
         cout << id + 1 << ' ';
     }
